% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twcv.R
\name{make_default_twcv_spec_set}
\alias{make_default_twcv_spec_set}
\title{Create a standard TWCV specification set}
\usage{
make_default_twcv_spec_set(
  predictor_vars,
  env_vars = predictor_vars,
  include_distance = TRUE,
  balance_by = 0.2,
  shrink_lambda = 0.2
)
}
\arguments{
\item{predictor_vars}{Character vector of focal predictor variables.}

\item{env_vars}{Character vector of environmental variables for an extended
specification.}

\item{include_distance}{Logical; include prediction distance \code{d}.}

\item{balance_by}{Quantile spacing used for discretization.}

\item{shrink_lambda}{Shrinkage parameter for calibration weights.}
}
\value{
Named list of TWCV specifications.
}
\description{
Create a standard TWCV specification set
}
