% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twcv.R
\name{normalize_weights}
\alias{normalize_weights}
\title{Normalize weights to unit mean}
\usage{
normalize_weights(w)
}
\arguments{
\item{w}{Numeric weight vector.}
}
\value{
Numeric vector with mean equal to one.
}
\description{
Normalize weights to unit mean
}
