% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.nndm}
\alias{plot.knndm}
\alias{plot.da_cv}
\title{Plot PDAV classes}
\usage{
\method{plot}{nndm}(x, type = "strict", stat = "ecdf", ...)

\method{plot}{knndm}(x, type = "strict", stat = "ecdf", ...)

\method{plot}{da_cv}(x, ...)
}
\arguments{
\item{x}{An object of type \emph{da_cv}.}

\item{type}{String, defaults to "strict" to show the original nearest neighbour distance definitions in the legend.
Alternatively, set to "simple" to have more intuitive labels.}

\item{stat}{String, defaults to "ecdf" but can be set to "density" to estimate density functions.}

\item{...}{other arguments.}
}
\description{
Generic plot function for prediction-domain adaptive validation results Classes
}
\author{
Carles Milà

Jan Linnenbrink
}
