% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twcv.R
\name{prepare_for_balancing}
\alias{prepare_for_balancing}
\title{Prepare variables for balancing via discretization}
\usage{
prepare_for_balancing(df, vars, ref_df, by = 0.2)
}
\arguments{
\item{df}{Data frame to transform.}

\item{vars}{Variables to transform.}

\item{ref_df}{Reference data frame.}

\item{by}{Quantile step size.}
}
\value{
Modified data frame with additional \code{*_cat} variables.
}
\description{
Transforms variables into categorical representations suitable for
balancing. Numeric variables are discretized using quantiles of a
reference distribution; categorical variables are aligned to the
reference support.
}
