% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdm_cv.R
\name{RDM_CV}
\alias{RDM_CV}
\title{Random k-fold CV (RDM-CV)}
\usage{
RDM_CV(samples, folds_k)
}
\arguments{
\item{samples}{sf object with POINT geometries. Must have a column "ID".}

\item{folds_k}{integer, number of folds}
}
\value{
Data frame with fold assignments
}
\description{
Random k-fold CV (RDM-CV)
}
