% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_cv_estimators.R
\name{rmse_metric}
\alias{rmse_metric}
\title{Root mean squared error metric}
\usage{
rmse_metric(obs, pred, weights = NULL)
}
\arguments{
\item{obs}{Numeric vector of observed values.}

\item{pred}{Numeric vector of predicted values.}

\item{weights}{Optional numeric weights.}
}
\value{
Numeric RMSE.
}
\description{
Root mean squared error metric
}
